Add IMAP auth and persist session as middleware
I'm handling the IMAP session as part of the Integrant system and then
passing down the session across the handlers using a middleware.

This might prove inefficient in the long run, I need to investigate how
to handle the connection better (maybe using a pool?).

This commit also adds Angus Mail dependency to handle auth and getting mails.