Workaround to trigger a render after API fetch
The problem occurs because the atom is updated inside a go block, which
prevents it from triggering a re-render.
A workaround is to wrap the atom update in a js/setTimeout with a delay
of 0 milliseconds. This schedules the update to occur in the next tick
of the event loop.

It's a horrible workaround and I don't want it to stay too long.